backend/rag/requirements.txt
flask
flask-cors
gemini_doc_query
vertexai
google-cloud
langchain-google-vertexai
langchain
langchain_community.document_loaders